refactor: 매직 넘버를 상수화
