+++ .eslintrc.cjs
@@ -14,7 +14,7 @@ module.exports = {
   rules: {
     'max-depth': ['error', 1],
     'max-params': ['error', 2],
-    'max-lines-per-function': ['error', { max: 10 }],
+    'max-lines-per-function': ['error', { max: 15 }],
     'import/extensions': 'off',
     'no-restricted-syntax': 'off',
     'no-await-in-loop': 'off',
